backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
django-import-export==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.2
sqlparse==0.4.3
Djoser==2.1.0
